Fix convertToSdkVisitor

+++ app/src/main/java/org/eyeseetea/malariacare/database/iomodules/dhis/exporter/ConvertToSDKVisitor.java
@@ -31,15 +31,19 @@ import org.eyeseetea.malariacare.R;
 import org.eyeseetea.malariacare.database.iomodules.dhis.importer.models.EventExtended;
 import org.eyeseetea.malariacare.database.model.CompositeScore;
 import org.eyeseetea.malariacare.database.model.ControlDataElement;
+import org.eyeseetea.malariacare.database.model.OrgUnit;
 import org.eyeseetea.malariacare.database.model.Survey;
+import org.eyeseetea.malariacare.database.model.TabGroup;
 import org.eyeseetea.malariacare.database.model.User;
 import org.eyeseetea.malariacare.database.model.Value;
 import org.eyeseetea.malariacare.database.utils.LocationMemory;
 import org.eyeseetea.malariacare.database.utils.PreferencesState;
 import org.eyeseetea.malariacare.database.utils.Session;
 import org.eyeseetea.malariacare.layout.score.ScoreRegister;
+import org.eyeseetea.malariacare.network.PullClient;
 import org.eyeseetea.malariacare.utils.Constants;
 import org.eyeseetea.malariacare.utils.Utils;
+
 import org.hisp.dhis.android.sdk.persistence.models.DataValue;
 import org.hisp.dhis.android.sdk.persistence.models.Event;
 import org.hisp.dhis.android.sdk.persistence.models.FailedItem;
@@ -66,16 +70,16 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
      */
     Context context;
 
-    String mainScoreCode;
-    String mainScoreACode;
-    String mainScoreBCode;
-    String mainScoreCCode;
-    String forwardOrderCode;
+    String mainScoreUID;
+    String mainScoreAUID;
+    String mainScoreBUID;
+    String mainScoreCUID;
+    String forwardOrderUID;
 
-    String createdOnCode;
-    String createdByCode;
-    String updatedDateCode;
-    String updatedUserCode;
+    String createdOnUID;
+    String createdByUID;
+    String updatedDateUID;
+    String updatedUserUid;
     /**
      * List of surveys that are going to be pushed
      */
@@ -102,30 +106,46 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
      */
     Date uploadedDate;
 
+    /**
+     * Used to control if the actual survey/event is new or update
+     */
+    boolean updateEvent;
+
+
     ConvertToSDKVisitor(Context context){
         this.context=context;
-        // FIXME: We should create a visitor to translate the ControlDataElement class
-        mainScoreCode = ControlDataElement.findControlDataElementUid(context.getString(R.string.main_score_code));
-        mainScoreACode = ControlDataElement.findControlDataElementUid(context.getString(R.string.main_score_a_code));
-        mainScoreBCode = ControlDataElement.findControlDataElementUid(context.getString(R.string.main_score_b_code));
-        mainScoreCCode = ControlDataElement.findControlDataElementUid(context.getString(R.string.main_score_c_code));
-        forwardOrderCode = ControlDataElement.findControlDataElementUid(context.getString(R.string.forward_order_code));
+        mainScoreUID=context.getString(R.string.main_score);
+        mainScoreAUID=context.getString(R.string.main_score_a);
+        mainScoreBUID=context.getString(R.string.main_score_b);
+        mainScoreCUID=context.getString(R.string.main_score_c);
+        forwardOrderUID=context.getString(R.string.forward_order);
 
-        createdOnCode = ControlDataElement.findControlDataElementUid(context.getString(R.string.created_on_code));
-        createdByCode = ControlDataElement.findControlDataElementUid(context.getString(R.string.created_by_code));
-        updatedDateCode =ControlDataElement.findControlDataElementUid(context.getString(R.string.upload_date_code));
-        updatedUserCode = ControlDataElement.findControlDataElementUid(context.getString(R.string.created_by_code));
+        createdOnUID =context.getString(R.string.createdOnUID);
+        createdByUID =context.getString(R.string.createdByUid);
+        updatedDateUID=context.getString(R.string.uploadedDateUID);
+        updatedUserUid=context.getString(R.string.createdByUid);
         surveys = new ArrayList<>();
         events = new ArrayList<>();
     }
 
     @Override
     public void visit(Survey survey) throws Exception{
+        updateEvent =false;
+
+        uploadedDate =new Date();
+
         //Turn survey into an event
         this.currentSurvey=survey;
-
         Log.d(TAG,String.format("Creating event for survey (%d) ...",survey.getId_survey()));
-        this.currentEvent=buildEvent();
+        Log.d(TAG,String.format("Creating event for survey (%s) ...", survey.toString()));
+
+        //if the event exist in the survey, it will be patched, else, created.
+        if(survey.getEventUid()!=null) {
+            buildUpgradeEvent(survey);
+        }
+        else
+            this.currentEvent = buildEvent();
+        Log.d(TAG,currentEvent.toString());
 
         //Calculates scores and update survey
         Log.d(TAG,"Registering scores...");
@@ -139,16 +159,43 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
         }
 
         //Turn question values into dataValues
-        Log.d(TAG, "Creating datavalues from questions...");
-        for(Value value:survey.getValues()){
-            value.accept(this);
+        Log.d(TAG, "Creating datavalues from questions... Values"+survey.getValues().size());
+
+
+        for(Value value : survey.getValues()) {
+            if(updateEvent) {
+                if (value.getUploadedDate().after(currentSurvey.getUploadedDate())) {
+                    value.accept(this);
+                    Log.d(TAG, "Value saved: " + value);
+                }
+            }
+            else
+                    value.accept(this);
         }
 
-        Log.d(TAG,"Saving control dataelements");
+        //Update all the dates after checks the new values
+        updateEventDates();
+        Log.d(TAG,"Saving dates in control dataelements");
+
+        buildDateControlDataElements(survey);
+
+        Log.d(TAG, "Creating datavalues from other stuff...");
         buildControlDataElements(survey);
 
         //Annotate both objects to update its state once the process is over
         annotateSurveyAndEvent();
+    }
+
+    private Event forceLastSurvey(Survey survey) {
+        //download last survey uid
+        PullClient pullClient = new PullClient((DashboardActivity) DashboardActivity.dashboardActivity);
+        PullClient.EventInfo eventInfo = pullClient.getLastEventUid(survey.getOrgUnit(), survey.getTabGroup());
+        if(!survey.getEventUid().equals(eventInfo.getEventUid())){
+            survey.setEventUid(eventInfo.getEventUid());
+        }
+        if(survey.getEvent()==null)
+            return buildFakeEvent(survey.getOrgUnit(),survey.getTabGroup(), eventInfo);
+        return survey.getEvent();
     }
 
     @Override
@@ -165,6 +212,15 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
 
     @Override
     public void visit(Value value) {
+        if(updateEvent){
+            if(value.getQuestion()==null) {
+                //The controlDataelements values don't have question. It should be ignored  in a upload event.
+                return;
+            }
+            if(value.getOption()==null){
+                return;
+            }
+        }
         DataValue dataValue=new DataValue();
         dataValue.setDataElement(value.getQuestion().getUid());
         dataValue.setLocalEventId(currentEvent.getLocalId());
@@ -183,19 +239,61 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
      * Builds an event from a survey
      * @return
      */
-    private Event buildEvent() throws Exception{
-        currentEvent=new Event();
-
+    private Event buildEvent()throws Exception{
+        currentEvent =new Event();
         currentEvent.setStatus(Event.STATUS_COMPLETED);
         currentEvent.setFromServer(false);
         currentEvent.setOrganisationUnitId(currentSurvey.getOrgUnit().getUid());
         currentEvent.setProgramId(currentSurvey.getTabGroup().getProgram().getUid());
         currentEvent.setProgramStageId(currentSurvey.getTabGroup().getUid());
         updateEventLocation();
-        updateEventDates();
         Log.d(TAG, "Saving event " + currentEvent.toString());
         currentEvent.save();
         return currentEvent;
+    }
+
+    /**
+     * Builds an fake event only to send the new DataValues.
+     * @return
+     */
+    public static Event buildFakeEvent(OrgUnit orgUnit, TabGroup tabGroup, PullClient.EventInfo eventInfo) {
+        //a false event was created to path the event datavalues
+        Log.d(TAG, "Recovering Event:" + eventInfo.getEventUid() + " not exist");
+        Log.d(TAG, "Creating fake event to upgrade one event in the server");
+        Event event = new Event();
+        event.setUid(eventInfo.getEventUid());
+        event.setLastUpdated(EventExtended.format(eventInfo.getEventDate(), EventExtended.DHIS2_DATE_FORMAT));
+        event.setEventDate(EventExtended.format(eventInfo.getEventDate(), EventExtended.DHIS2_DATE_FORMAT));
+        event.setOrganisationUnitId(orgUnit.getUid());
+        event.setProgramId(tabGroup.getProgram().getUid());
+        event.setProgramStageId(tabGroup.getUid());
+        event.save();
+        return event;
+    }
+
+    /**
+     * Builds or modify a event to be upgraded.
+     * @return
+     */
+    private void buildUpgradeEvent(Survey survey) {
+        updateEvent =true;
+        currentEvent = forceLastSurvey(survey);
+
+        if(currentEvent != null){
+            currentEvent.setCreated(null);
+            uploadedDate = currentSurvey.getUploadedDate();
+            String date=EventExtended.format(currentSurvey.getCompletionDate(), EventExtended.DHIS2_DATE_FORMAT);
+            currentEvent.setEventDate(date);
+            currentEvent.setLastUpdated(EventExtended.format(currentSurvey.getUploadedDate(), EventExtended.DHIS2_DATE_FORMAT));
+            //It's necesary, Set"from server" as false to upload the event
+            currentEvent.setFromServer(false);
+            currentEvent.setStatus(Event.STATUS_COMPLETED);
+            currentEvent.save();
+        }
+        else
+        {
+            Log.d(TAG, "Error Creating/Recovering Event:"+survey.getEventUid()+" not exist");
+        }
     }
 
     /**
@@ -203,16 +301,13 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
      */
     private void updateEventDates() {
 
-        //Sent date 'now' (this change will be saves after successful push)
-        currentSurvey.setUploadedDate(new Date());
-
-        uploadedDate =currentSurvey.getUploadedDate();
-
         // NOTE: do not try to set the event creation date. SDK will try to update the event in the next push instead of creating it and that will crash
-        currentEvent.setEventDate(EventExtended.format(currentSurvey.getCompletionDate(), EventExtended.DHIS2_DATE_FORMAT));
-        currentEvent.setDueDate(EventExtended.format(currentSurvey.getScheduledDate(),EventExtended.DHIS2_DATE_FORMAT));
+        String date=EventExtended.format(currentSurvey.getCompletionDate(), EventExtended.DHIS2_DATE_FORMAT);
+        currentEvent.setEventDate(date);
+        currentEvent.setDueDate(EventExtended.format(currentSurvey.getScheduledDate(), EventExtended.DHIS2_DATE_FORMAT));
         //Not used
-        currentEvent.setLastUpdated(EventExtended.format(currentSurvey.getUploadedDate(),EventExtended.DHIS2_DATE_FORMAT));
+        currentEvent.setLastUpdated(EventExtended.format(currentSurvey.getUploadedDate(), EventExtended.DHIS2_DATE_FORMAT));
+        currentEvent.save();
         }
 
     /**
@@ -221,38 +316,38 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
      */
     private void buildControlDataElements(Survey survey) {
 
-        //It Checks if the dataelement exists, before build and save the datavalue
+        //MainScoreUID
+        buildAndSaveDataValue(mainScoreUID, survey.getType());
+
+        //MainScore A
+        buildAndSaveDataValue(mainScoreAUID, survey.isTypeA() ? "true" : "false");
+
+        //MainScore B
+        buildAndSaveDataValue(mainScoreBUID, survey.isTypeB() ? "true" : "false");
+
+        //MainScoreC
+        buildAndSaveDataValue(mainScoreCUID, survey.isTypeC() ? "true" : "false");
+
+        //Forward Order
+        buildAndSaveDataValue(forwardOrderUID, context.getString(R.string.forward_order_value));
+    }
+
+    /**
+     * Builds several datavalues from the Dates of the survey
+     * @param survey
+     */
+    private void buildDateControlDataElements(Survey survey) {
         //Created date
-        if(createdOnCode!=null && !createdByCode.equals(""))
-            buildAndSaveDataValue(createdOnCode, EventExtended.format(survey.getCreationDate(), EventExtended.AMERICAN_DATE_FORMAT));
+        buildAndSaveDataValue(createdOnUID, EventExtended.format(survey.getCreationDate(), EventExtended.AMERICAN_DATE_FORMAT));
 
         //Updated date
-        if(updatedDateCode!=null && !updatedDateCode.equals(""))
-            buildAndSaveDataValue(updatedDateCode, EventExtended.format(survey.getUploadedDate(), EventExtended.AMERICAN_DATE_FORMAT));
+        buildAndSaveDataValue(updatedDateUID, EventExtended.format(uploadedDate, EventExtended.AMERICAN_DATE_FORMAT));
 
         //Updated by user
-        if(updatedUserCode!=null && !updatedUserCode.equals(""))
-            buildAndSaveDataValue(updatedUserCode, Session.getUser().getUid());
+        buildAndSaveDataValue(updatedUserUid, Session.getUser().getUid());
 
         //Updated by user
-        if(createdByCode!=null && !createdByCode.equals(""))
-            buildAndSaveDataValue(createdByCode, Session.getUser().getUid());
-
-        //MainScoreUID
-        if(mainScoreCode!=null)
-            buildAndSaveDataValue(mainScoreCode, survey.getType());
-
-        if(mainScoreACode!=null)
-            buildAndSaveDataValue(mainScoreACode, survey.isTypeA() ? "true" : "false");
-
-        if(mainScoreBCode!=null)
-            buildAndSaveDataValue(mainScoreBCode, survey.isTypeB() ? "true" : "false");
-
-        if(mainScoreCCode!=null)
-            buildAndSaveDataValue(mainScoreCCode, survey.isTypeC() ? "true" : "false");
-
-        if(forwardOrderCode!=null)
-            buildAndSaveDataValue(forwardOrderCode, context.getString(R.string.forward_order_value));
+        buildAndSaveDataValue(createdByUID, Session.getUser().getUid());
     }
 
     private void buildAndSaveDataValue(String UID, String value){
@@ -274,7 +369,6 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
     private void updateSurvey(List<CompositeScore> compositeScores){
         currentSurvey.setMainScore(ScoreRegister.calculateMainScore(compositeScores));
         currentSurvey.setStatus(Constants.SURVEY_SENT);
-        currentSurvey.setUploadedDate(uploadedDate);
         currentSurvey.setEventUid(currentEvent.getUid());
     }
 
@@ -304,6 +398,7 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
      */
     private void annotateSurveyAndEvent() {
         surveys.add(currentSurvey);
+        currentEvent.setLastUpdated(EventExtended.format(uploadedDate, EventExtended.DHIS2_DATE_FORMAT));
         events.add(currentEvent);
         Log.d(TAG, String.format("%d surveys converted so far", surveys.size()));
     }
@@ -316,8 +411,9 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
             Survey iSurvey=surveys.get(i);
             Event iEvent=events.get(i);
             ImportSummary importSummary=importSummaryMap.get(iEvent.getLocalId());
-            FailedItem failedItem= hasConflict(iEvent.getLocalId());
+            FailedItem failedItem= EventExtended.hasConflict(iEvent.getLocalId());
             if(hasImportSummaryErrors(importSummary) || failedItem!=null){
+                Log.d(TAG, importSummary.toString());
                 //Some error happened -> move back to completed
                 if(failedItem!=null) {
                     iSurvey.setStatus(Constants.SURVEY_COMPLETED);
@@ -332,34 +428,32 @@ public class ConvertToSDKVisitor implements IConvertToSDKVisitor {
                 }
                 iSurvey.save();
 
-                //Generated event must be remove too
-                iEvent.delete();
-                Log.d(TAG, "PUSH process...Fail pushing survey: " + iSurvey.getId_survey());
+                //Generated event must be remove too (or saved in a modify import summary)
+                if(iSurvey.getStatus()!=Constants.SURVEY_CONFLICT && importSummary.getStatus().equals("SUCCESS")) {
+                    saveSurveyFromImportSummary(iSurvey, iEvent);
+                    Log.d(TAG, "PUSH process...Survey uploaded: " + iSurvey.getId_survey());
+                }
+                else{
+                    Log.d(TAG, "PUSH process...Fail pushing survey: " + iSurvey.getId_survey());
+                    iEvent.delete();
+                }
             }else{
-                iSurvey.setStatus(Constants.SURVEY_SENT);
-                iSurvey.saveMainScore();
-                iSurvey.save();
-
-                //To avoid several pushes
-                iEvent.setFromServer(true);
-                iEvent.save();
-
-                Log.d(TAG, "PUSH process...OK. Survey and Event saved");
+                saveSurveyFromImportSummary(iSurvey, iEvent);
             }
         }
     }
 
-    /**
-     * Checks whether the given event contains errors in SDK FailedItem table or has been successful.
-     * If not return null, it is becouse this item had a conflict.
-     * @param localId
-     * @return
-     */
-    private FailedItem hasConflict(long localId){
-        return  new Select()
-                        .from(FailedItem.class)
-                        .where(Condition.column(FailedItem$Table.ITEMID)
-                                .is(localId)).querySingle();
+    private void saveSurveyFromImportSummary(Survey iSurvey, Event iEvent) {
+        iSurvey.setStatus(Constants.SURVEY_SENT);
+        iSurvey.setUploadedDate(uploadedDate);
+        iSurvey.saveMainScore();
+        iSurvey.save();
+
+        //To avoid several pushes
+        iEvent.setFromServer(true);
+        iEvent.save();
+
+        Log.d(TAG, "PUSH process...OK. Survey and Event saved");
     }
 
     /**
